refactor db_client
