+++ src/identify_updates_of_topics/_utils/database.py
@@ -5,11 +5,9 @@ from functools import lru_cache
 from typing import Any
 
 import sqlalchemy
-from sqlalchemy.engine import Connection
-from sqlalchemy.engine.base import Engine
 
 from _dependencies.commons import sqlalchemy_get_pool
-from _dependencies.pubsub import notify_admin
+from _dependencies.db_client import DBClientBase, DBKeyValueStorageMixin
 
 from .topics_commons import (
     ChangeLogLine,
@@ -19,13 +17,7 @@ from .topics_commons import (
 )
 
 
-class DBClient:
-    def __init__(self, db: Engine) -> None:
-        self._db = db
-
-    def connect(self) -> Connection:
-        return self._db.connect()
-
+class DBClient(DBClientBase,DBKeyValueStorageMixin):
     def get_the_list_of_ignored_folders(self) -> list[int]:
         """get the list of folders which does not contain searches – thus should be ignored"""
 
@@ -411,32 +403,6 @@ class DBClient:
                                     """)
 
             return conn.execute(stmt).fetchall()
-
-    def get_key_value_item(self, key: str) -> Any:
-        with self.connect() as conn:
-            stmt = sqlalchemy.text("""
-                SELECT value FROM key_value_storage WHERE key=:key;
-                                   """)
-            raw_data = conn.execute(stmt, key=key).fetchone()
-            return raw_data[0] if raw_data else None
-
-    def set_key_value_item(self, key: str, value: Any) -> None:
-        with self.connect() as conn:
-            stmt = sqlalchemy.text("""
-                INSERT INTO key_value_storage 
-                (key, value) 
-                VALUES (:key, :value) 
-                ON CONFLICT (key) DO UPDATE SET value = :value ; 
-                                   """)
-            conn.execute(stmt, key=key, value=json.dumps(value))
-
-    def delete_key_value_item(self, key: str) -> None:
-        with self.connect() as conn:
-            stmt = sqlalchemy.text("""
-                DELETE FROM key_value_storage 
-                WHERE key=:key; 
-                                   """)
-            conn.execute(stmt, key=key)
 
 
 @lru_cache
