+++ src/send_notifications/main.py
@@ -6,15 +6,14 @@ import logging
 import re
 import time
 from concurrent.futures import ThreadPoolExecutor, wait
-from contextlib import contextmanager
 from dataclasses import dataclass, field
 from functools import lru_cache
-from typing import Any, Iterator, List
+from typing import Any
 
 import sqlalchemy
-from sqlalchemy.engine.base import Connection, Engine
 
 from _dependencies.commons import setup_logging, sqlalchemy_get_pool
+from _dependencies.db_client import DBClientBase
 from _dependencies.lock_manager import FunctionLockError, lock_manager
 from _dependencies.misc import generate_random_function_id, tg_api_main_account
 from _dependencies.pubsub import Ctx, notify_admin, pubsub_send_notifications
@@ -62,18 +61,10 @@ class MessageToSend:
 
 @lru_cache
 def db() -> 'DBClient':
-    return DBClient(_pool=sqlalchemy_get_pool())
+    return DBClient(db=sqlalchemy_get_pool())
 
 
-@dataclass
-class DBClient:
-    _pool: Engine
-
-    @contextmanager
-    def connect(self) -> Iterator[Connection]:
-        with self._pool.connect() as connection:
-            yield connection
-
+class DBClient(DBClientBase):
     def get_notifs_to_send(self, select_doubling: bool) -> list['MessageToSend']:
         """return notifications which should be sent"""
         with self.connect() as conn:
@@ -430,7 +421,7 @@ def _process_logs_with_completed_sending(
 
 def finish_time_analytics(
     time_analytics: TimeAnalytics,
-    list_of_change_ids: List,
+    list_of_change_ids: list,
 ) -> None:
     """Make final steps for time analytics: inform admin, log, record statistics into PSQL"""
